Use case 1 til merge (rigtig)
