+++ ST2PRJ2_F17/ST2PRJ2_F17.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26403.7
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ST2PRJ2_F17", "ST2PRJ2_F17\ST2PRJ2_F17.csproj", "{992BCFCE-BF11-4EA2-A6B7-9A6AD14630AE}"
 EndProject
