+++ ST2PRJ2_F17/ST2PRJ2_F17/ST2PRJ2_F17.csproj
@@ -52,7 +52,7 @@
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
-    <Compile Include="Logiklag.cs" />
+    <Compile Include="Log_ind_controller.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Form1.resx">
